+++ crates/adapters/tardis/src/lib.rs
@@ -33,13 +33,14 @@
 //! for the [nautilus_trader](https://pypi.org/project/nautilus_trader) Python package,
 //! or as part of a Rust only build.
 //!
+//! - `arrow`: Enables Apache Arrow and Parquet data support.
 //! - `examples`: Enables the crate's example binaries.
 //! - `extension-module`: Builds as a Python extension module.
 //! - `high-precision` (default): Enables
 //!   [high-precision mode](https://nautilustrader.io/docs/nightly/getting_started/installation/#precision-mode)
 //!   to use 128-bit value types.
 //! - `python`: Enables Python bindings from [PyO3](https://pyo3.rs).
-//! - `replay` (default): Enables market data replay functionality.
+//! - `replay`: Enables market data replay functionality (implies `arrow`).
 
 #![warn(rustc::all)]
 #![deny(unsafe_code)]